pyproject.toml
[tool.poetry]
name = "terraform-worker"
version = "0.12.0"
description = "An orchestration tool for Terraform"
authors = ["Richard Maynard <richard.maynard@gmail.com>"]
packages = [
    { include="tfworker", from="." },
]
readme = "README.md"

repository = "https://github.com/ephur/terraform-worker"
homepage = "https://github.com/ephur/terraform-worker"
documentation = "https://github.com/ephur/terraform-worker"

license = "Apache-2.0"

classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.8",
    "Topic :: System :: Systems Administration"
]

[tool.poetry.dependencies]
python = "^3.11"
atlassian-python-api = "^3.41"
boto3 = "^1.34"
click = "^8.1"
google-cloud-storage = "^2.17"
jinja2 = "^3.1"
mergedeep = "^1.3"
pydantic = "^2.7"
python-hcl2 = "^4.3"
pyyaml = "^6.0"
setuptools = "^70.0"
pydantic-settings = "^2.3.4"
packaging = "^24.1"

[tool.poetry.scripts]
worker = 'tfworker.cli:cli'

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^24.4"
coverage = "^7.5"
deepdiff = "^7.0"
flake8 = "^7.0"
ipython = "^8.24"
isort = "^5.13"
pytest = "^8.2"
pytest-cov = "^5.0"
pytest-depends = "^1.0"
pytest-mock = "^3.14"
pytest-socket = "^0.7"
pytest-timeout = "2.3.1"
seed-isort-config = "^2.2"
Sphinx = "^7.3"
wheel = "^0.43"
moto = {version = "^5.0.10", extras = ["sts", "dynamodb", "s3"]}


[tool.pytest.ini_options]
addopts = "--capture=sys --cov=tfworker --cov-report= -m 'not performance'"
markers = [
    "performance: mark a test as a performance test"
]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
